requirements.txt
# local package
-e .

# external requirements
click
coverage
flake8
python-dotenv>=0.5.1
wandb
hydra-core
h5py
pandas
numpy
rdkit-pypi
scipy
torch
torch-scatter
torch-sparse
torch-geometric

# you must install pytables with conda

